refactor(gateway): extract http_errors.go from server.go
Phase 3b file split: moves writeJSONError, getErrorTypeAndTitle,
writeValidationError, and writeInternalError (RFC 7807 error response
helpers) into their own file. Pure structural move: 154 Ginkgo specs
still pass.
